Create and export getSearchErrorOverrideDisplay from search_interceptor
