+++ src/plugins/data/public/search/index.ts
@@ -57,6 +57,7 @@ export { getEsPreference } from './es_search';
 
 export type { SearchInterceptorDeps } from './search_interceptor';
 export { SearchInterceptor } from './search_interceptor';
+export { getSearchErrorOverrideDisplay } from './search_interceptor/utils';
 export * from './errors';
 
 export { SearchService } from './search_service';
